Create CardsController and rename some files
