+++ app/assets/javascripts/cards.js
