+++ app/assets/stylesheets/application.scss
@@ -18,8 +18,8 @@
 @import 'registration_info';
 // @import 'confirm_tel';
 @import 'register_user_shipping';
-@import 'users_add_card';
-@import 'users_add_card_info';
+@import 'cards/show';
+@import 'cards/new';
 @import 'register_complete';
 @import 'edit_user_info';
 @import 'app_banner';
