+++ app/assets/stylesheets/cards/_new.scss
