+++ app/assets/stylesheets/cards/_show.scss
