+++ app/controllers/users_controller.rb
@@ -1,12 +1,8 @@
 class UsersController < ApplicationController
-  before_action :authenticate_user!, only: [:add_card, :add_card_info, :pay, :delete_card]
-  before_action :prepare_payjp, only: [:add_card, :add_card_info, :pay, :delete_card]
 
   def index
   end
 
-
-# -----------------ユーザー登録処理-----------------------
   def register_user_tel
     session[:user_info] = params[:user]
   end
@@ -27,27 +23,6 @@ class UsersController < ApplicationController
   end
 
 
-# --------------クレジットカード追加処理--------------------
-  def add_card
-  end
-
-  def add_card_info
-    redirect_to action: :add_card if @card
-  end
-
-  def pay
-    token = params[:token]
-    @customer.cards.create(card: token) if @customer.cards.count == 0
-    redirect_to action: :add_card, notice: "カードが登録されました。"
-  end
-
-  def delete_card
-    @card.delete
-    redirect_to action: :add_card, notice: "カード情報が削除されました。"
-  end
-
-
-# --------------プライベートメソッド--------------------
   private
   # def user_info_params
   #   params.permit(:email,:nickname, :family_name, :first_name, :family_name_phonetic, :first_name_phonetic, :birth_year, :birth_month, :birth_day)
@@ -61,14 +36,5 @@ class UsersController < ApplicationController
     params.require(:user).permit()
   end
 
-  def prepare_payjp
-    Payjp.api_key = ENV["PAYJP_SK_TEST"]
-    unless current_user.payjp_id
-      current_user.payjp_id = Payjp::Customer.create(description: 'test').id
-      current_user.save
-    end
-    @customer = Payjp::Customer.retrieve(current_user.payjp_id)
-    @card = @customer.cards.retrieve(@customer.default_card) if @customer.cards.count > 0
-  end
 
 end
