+++ app/helpers/pay_helper.rb
@@ -1,0 +1,2 @@
+module PayHelper
+end
