+++ app/views/cards/show.html.haml
@@ -29,9 +29,9 @@
                   ************#{@card.last4}
                 .add-card__main__content__info--exp
                   #{@card.exp_month}/#{@card.exp_year}
-                = link_to '削除する', delete_card_users_path, method: :delete, class: "delete-button"
+                = link_to '削除する', card_path, method: :delete, class: "delete-button"
             - else
-              = link_to add_card_info_users_path, method: :get, class: "add-button" do
+              = link_to new_card_path, method: :get, class: "add-button" do
                 =fa_icon('credit-card')
                 クレジットカードを追加する
           .add-card__footer
